lambda_functions/gather_nutrition_data_lambda/lambda_environment/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.54.1
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==11.0.0
psycopg2-binary==2.9.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
